+++ CHANGELOG.md
@@ -370,3 +370,4 @@
 * `modify` and `replaces` methods of `State` effect returns result of an applied function
 * Change precedence of `->>` method of `Traversable` - from 3 to 5
 * Rename `Stack` datastructure to `List`
+* Define `Stack` interface with `push` and `pop` methods